InferenceService URL as target host of Model VS (#23)
* InferenceService URL as target host of Model VS

Previously, we use InferenceService Predictor URL for Model Endpoint VirtualService target host when serving traffic to model version.
E.g., when we serve model version 1, we will have this url (model-version-1-predictor-default.domain.com) as host.

The problem would arise if we deploy a model with a transformer. If we keep using Predictor URL as Host, we will skip sending request to Transformer. To fix it, we can do one of these:
1. Check if transformer exist, then use Transformer URL as Model Endpoint VirtualService host, or
2. Use InferenceService Host URL

This commit will use the option #2.

* Add test serving transformer + model