Make sdk compatible with python 3.7 to 3.10 (#293)
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. Run unit tests and ensure that they are passing
2. If your change introduces any API changes, make sure to update the
e2e tests
3. Make sure documentation is updated for your PR!

-->

**What this PR does / why we need it**:
<!-- Explain here the context and why you're making the change. What is
the problem you're trying to solve. --->
As python 3.7 is reaching the end of life, we are upgrading merlin to
support development in python 3.8, 3.9 and 3.10.
In the first part of this upgrade, we tested our sdk against the various
python versions and updated the merlin examples accordingly.

**Caveat**: This PR only makes sdk installable, usable and runnable
using python 3.7 to 3.10. Subsequent work will need to be carried out to
make Merlin fully compatible with python 3.8 to 3.10

**Test results:** 
1. No change in current setup dependencies are required for merlin-sdk
to be installed on python 3.7 to 3.10.
2. We will need to update scikit-learn to >=1.1.2 and joblib to >=1.2.0
to fully support development work on python 3.8 to 3.10. However, this
will be done after we update our kserve to use the respective libraries
versions.
3. In order to support python 3.11, we will need to wait for new release
of scipy. On top of that, protobuf will need to be >4 and mlflow
>=1.26.1. As we still largely depend on libraries using protobuf 3, this
version will not support python 3.11 at this stage.

**Subsequent work:**
To make Merlin fully compatible with python 3.8 to 3.10, the following
tasks need to be completed subsequently:
1. We will need to update scikit-learn to >=1.1.2 and joblib to >=1.2.0
to fully support development work on python 3.8 to 3.10. However, this
will be done after we update our kserve to use the respective libraries
versions.
2. Update pyfunc server and batch prediction server to build and run
pyfunc images according to python version used.

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
Updated dependencies version to use some newer libraries for setup
Updated merlin examples to be compatible with python 3.7 to 3.10
```

**Checklist**

- [ ] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [x] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes