Upgrade pyfunc & batch predictor to support python3.7 to 3.10 (#301)
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. Run unit tests and ensure that they are passing
2. If your change introduces any API changes, make sure to update the
e2e tests
3. Make sure documentation is updated for your PR!

-->

**What this PR does / why we need it**:
Python 3.7 will become obsolete soon. We need to upgrade Merlin to
support newer python versions, namely python 3.8, 3.9 and 3.10.

As part 2 & 3 of this effort, this PR mainly does the following:
1. Upgrade pyfunc to support python 3.7 to 3.10
2. Upgrade batch predictor to support python 3.7 to 3.10

**Notable changes:**
1. Create base image for each python version for pyfunc*: i.e.
merlin-pyfunc-base-py37 etc
2. Create base images for each python version for batch*: i.e.
merlin-pyspark-base-py37 etc
3. Update Gitlab pipeline to test sdk, api, pyfunc and batch using
python 3.7 to 3.10
4. Capture user's python version major and minor when user create a
model version. This is be used to determine the python version of the
model to be logged/deployed. Any python version settings in the env.yaml
file will be replaced to the captured version. In case user is still
using older merlin sdk, python 3.7.* is assumed to ensure backward
compatibility.
5. Mapping of python version captured from user to respective
pyfunc/batch base docker image.

**Base images Updates:**
1. For batch base Dockerfile, pyspark 2.4.5 is not compatible with >=
py3.8, hence upgraded till 3.0.0 which does not officially support py3.9
and py3.10, but tested to work on our servers. We are not able to
upgrade to 3.1.1 as it is not compatible with our spark operator version
(completed jobs won't shutdown)
2. miniconda has been replaced with miniforge as latest miniconda
available does not support py3.10 as of now.


**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
- Support for python 3.7 to 3.10 for Merlin (SDK, Pyfunc and Batch predictors)
- Backward compatibility with older sdk
```

**Checklist**

- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [x] Updated documentation
- [x] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes