Make API server use its environment label in all deployments (#370)
**What this PR does / why we need it**:
Kubernetes resources deployed by the Merlin API server (Merlin services,
batch jobs, etc.) are currently given the resource label `[label-prefix]
+ "environment": [environment name]`. This environment name is obtained
from the version endpoint's environment name (which reflects the name of
the cluster in which the resource is deployed in).

However, this is inconsistent from the intended value of the label,
which should instead be the same as the API server's
[environment](https://github.com/gojek/merlin/blob/main/api/config/config.go#L38),
i.e. `[label-prefix] + "environment": [environment name found in the API
server's config]`. As such this PR introduces some minor refactoring to
remove change the said environment name, as well as some other minor
refactoring of test case variables.

**Which issue(s) this PR fixes**:
Fixes #

**Does this PR introduce a user-facing change?**:
```release-note
NONE
```

**Checklist**

- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes