Disallow changing deployment mode for pending/running endpoints (#377)
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. Run unit tests and ensure that they are passing
2. If your change introduces any API changes, make sure to update the
e2e tests
3. Make sure documentation is updated for your PR!

-->

**What this PR does / why we need it**:
<!-- Explain here the context and why you're making the change. What is
the problem you're trying to solve. --->

This PR fixes an issue with switching the deployment mode of a model
version, where the K8s resources are not cleaned up properly.

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

When switching the deployment mode of an existing (running) deployment
from serverless -> raw or vice versa, Kserve uses slightly different
naming conventions for the underlying deployments and they are not
cleaned up properly. Eg:

```
NAME                                                        READY   UP-TO-DATE   AVAILABLE   AGE
tensorflow-sample-s-27-predictor-default                    1/1     1            1           50s
tensorflow-sample-s-27-predictor-default-00001-deployment   1/1     1            1           97s
```

We are already preventing changing the deployment mode of a serving
endpoint because it cannot be done gracefully. The same argument can be
extended to a running endpoint as well. In addition, the issue with the
deployment resource not being cleaned up can be circumvented in the same
way. Thus, in all, we now block "updating" an endpoint in any of the 3
states:
* Pending
* Running
* Serving

Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
Changed: Deployment mode of model endpoints that are in pending / running state can no longer be changed. Previously, this rules was only applicable to model endpoints in serving state.
```

**Checklist**

- [x] Added unit test, integration, and/or e2e tests
- [ ] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes

---------

Co-authored-by: Krithika Sundararajan <krithika.sundararajan@go-jek.com>