Improve Image Builder Jobs Label and Annotation (#380)
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. Run unit tests and ensure that they are passing
2. If your change introduces any API changes, make sure to update the
e2e tests
3. Make sure documentation is updated for your PR!

-->

**What this PR does / why we need it**:
<!-- Explain here the context and why you're making the change. What is
the problem you're trying to solve. --->

1. The image-building jobs in Merlin are timing out. After some
investigation, we found that one of the root causes is the node pool got
scaled down resulting in the image building pods to be rescheduled. This
PR adds "cluster-autoscaler.kubernetes.io/safe-to-evict": "false" to
avoid the pod get killed and rescheduled.
2. In Refactor deployed model labels #346, we managed to propagate user
labels to models and batch prediction jobs, but not to image builder
jobs (here). This PR includes adding project + model version labels to
image builder job.
3. This PR also improve e2e test by:
a. Removing the duplicate of kserve-controller-manager statefulset as in
kserve 0.9.0 manifest has both of statefulset and deployment
  b. Removing kserve-controller-manager's cpu limits

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
NONE
```

**Checklist**

- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [x] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes