chore: Refactor error logging from API handlers (#396)
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. Run unit tests and ensure that they are passing
2. If your change introduces any API changes, make sure to update the
e2e tests
3. Make sure documentation is updated for your PR!

-->

**What this PR does / why we need it**:
<!-- Explain here the context and why you're making the change. What is
the problem you're trying to solve. --->

This PR refactors the errors logged by the API handlers to do the
following:
* `api/utils/stacktrace.go` - Utility function for capturing stacktrace
(largely inspired by the examples in the [official
docs](https://pkg.go.dev/runtime#Frames))
* `api/api/response.go` - Add `stacktrace string` to the `Response`
struct. All utility functions for non-successful responses will capture
the stacktrace (which is closer to the origin of the error) in the
response.
* `api/api/router.go` - The handler function will log an error to the
console will relevant details about the API request / response (such as
the route name, error response and stacktrace), when the response
indicates failure. The `stacktrace` in the log will be taken from the
`Response` data and overrides the default stacktrace printed by the
`zap` logger (which would not have been of use in this case, as it
points to the logging line in the `HandlerFunc`).

<img width="1428" alt="Screenshot 2023-06-06 at 9 44 38 AM"
src="https://github.com/caraml-dev/merlin/assets/23465343/46744daf-3f58-468a-9da9-85e92ef5709f">

* `api/` - The following changes are made to the error messages and
logging.
- All error messages in API responses now include the original error
received from the service layer. On the other hand, some info that can
be inferred from the request has been removed from the response (eg: For
a non-existent model version 2 under a model with id=1, GET
`/models/1/versions/2` will return `Model version not found` as opposed
to `Model version 2 not found`). Besides this, there is a general clean
up of error messages to make them more meaningful (eg: previously we
were showing errors like `Model version 1 for version 2`; now it is
`Model version not found`).
- Individual log statements for the errors to the console have been
removed (as the router is expected to handle this for all failures, via
the wrapper `HandlerFunc`). The only exception is `api/api/log_api.go`
which is added as a raw handler to the router.

**Which issue(s) this PR fixes**:
<!--
*Automatically closes linked issue when PR is merged.
Usage: `Fixes #<issue number>`, or `Fixes (paste link of issue)`.
-->

The additional error details are meant to help with debugging API
errors.

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
Propagate original error in the response, back to the caller, for non 2XX responses.
```

**Checklist**

- [ ] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes