Bugfix: Add image building job deletion delay (#411)
**What this PR does / why we need it**:
As of present, when the Merlin API server triggers an image building job
as part of launching a batch prediction job/service, it checks for the
presence of an existing image building job in the cluster with the same
name. The status of the image building job found (if it is found),
corresponds to a variety of situations:
- Found; running -> an existing image building job is already taking
place and still running -> wait for the job to complete
- Found; failed -> an existing image building job has failed -> delete
the old job -> begin a new one
- Not found -> begin a new one

In particular, we're interested in the **_second_** scenario, where
there a new job is triggered right after the deletion of the old
(failed) run (see
[this](https://github.com/caraml-dev/merlin/blob/9d2689df94ba0f622ce01bdb60ed27a9af37a603/api/pkg/imagebuilder/imagebuilder.go#L142)
for more details):

```go
err = jobClient.Delete(ctx, job.Name, metav1.DeleteOptions{})
if err != nil {
	log.Errorf("error deleting job: %v", err)
return "", ErrDeleteFailedJob
}

jobSpec, err := c.createKanikoJobSpec(project, model, version)
if err != nil {
	log.Errorf("unable to create job spec %s, error: %v", imageRef, err)
	return "", ErrUnableToCreateJobSpec{
		Message: err.Error(),
	}
}

job, err = jobClient.Create(ctx, jobSpec, metav1.CreateOptions{})
if err != nil {
	log.Errorf("unable to build image %s, error: %v", imageRef, err)
	return "", ErrUnableToBuildImage{
		Message: err.Error(),
	}
}
```

A potential problem can occur whereby the creation of the new job
happens _too quickly_ after the deletion of the job, due to the k8s
resources having insufficient time to be removed from the cluster (the
kube API server would return an error saying that the creation of the
new job has failed due to the existence of another job with the very
same name). When this happens, the (new) image building job fails
immediately.

This PR thus introduces a new method to wait for the deletion to be
complete before proceeding with the creation of another image builder
job.

#### Main Modifications
- `api/pkg/imagebuilder/imagebuilder.go` - Addition of an additional
helper method to wait for image building jobs to be completely removed
from the cluster

**Which issue(s) this PR fixes**:
Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
NONE
```

**Checklist**

- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes