Feat/delete api (#389)
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. Run unit tests and ensure that they are passing
2. If your change introduces any API changes, make sure to update the
e2e tests
3. Make sure documentation is updated for your PR!

-->

**What this PR does / why we need it**: Allowing User to delete model
and model version
## Delete Feature
This PR contain the API to allow user to delete model, model version and
prediction job.
Deleting an model version have several constraint that must be met,
including:
- For `pyfunc_v2` model 
  - No `pending` or `running` prediction jobs that use the model version
- No `pending`, `running`, or `serving` endpoints that use the model
version
- For non `pyfunc_v2` model 
- No `pending`, `running`, or `serving` endpoints that use the model
version

If the model version can be deleted, this feature will also delete other
related entities such as:
- For `pyfunc_v2` model
- Prediction job with status `failed`, `failed_submission`, `completed`,
and `terminated`
  - Endpoint with status `failed`, `terminated`
- For non `pyfunc_v2`
  - Endpoint with status `failed`, `terminated`

A popup will appear showing the entity that is blocking the deletion
process (if the constraints have not been met) and the entities that
will be deleted if the constraints are met. Even though that the API is
facing some changes on checking the endpoint for the `pyfunc_v2` model,
the ui wont check the endpoint status, since current implementation
prevent the user for deploying the `pyfunc_v2` model on the frontend.

## Delete Prediction Jobs
This PR change on how the implementation of delete prediction jobs work,
instead of changing the status to `terminated` on `StopJob`, now the
`StopPredictionJob` function will also delete the job data from the
`spark` and from the `database`. If the job **doesn't have spark
application**, then instead of throwing an error, the deletion process
will proceed with deleting the data from the database
<!-- Explain here the context and why you're making the change. What is
the problem you're trying to solve. --->


**Does this PR introduce a user-facing change?**: No
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
NONE
```

**Checklist**

- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [x] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes

---------

Co-authored-by: Alexander <alexander.i.interns@aux.gojek.com>