Convert API configs from env vars to config file values (#397)
**What this PR does / why we need it**:
In line with the general direction of moving away from using environment
variables to pass configuration to applications, to using configuration
files (.yaml), the Merlin API server has been refactored in order to
read its configuration values off a specified file instead of
environment variables. This change also requires an accompanying change
to its Helm chart, as it would no longer be able to utilise the same
environment variables to derive its configuration. See
https://github.com/caraml-dev/helm-charts/pull/254 for the accompanying
PR for this change (that PR has to be merged first before this PR).

**Which issue(s) this PR fixes**:
Most of this PR involves removing the `envconfig` tags from the various
fields in `api/config/config.go`. A couple of additional helper methods
(copied over from Turing) are introduced in order to perform the parsing
of the configs from a `.yaml` config file.

~~NOTE that the e2e tests have been temporarily been commented out to
allow the publishing of a Merlin API server image to allow the changes
in the Helm chart repo to pass its CICD tests and to get merged.~~

**Does this PR introduce a user-facing change?**:
NONE

```release-note
NONE
```

**Checklist**

- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes