Set initial scale during redeployment of a model version (#431)
<!--  Thanks for sending a pull request!  Here are some tips for you:

1. Run unit tests and ensure that they are passing
2. If your change introduces any API changes, make sure to update the
e2e tests
3. Make sure documentation is updated for your PR!

-->

**What this PR does / why we need it**:
<!-- Explain here the context and why you're making the change. What is
the problem you're trying to solve. --->

This PR implements setting the initial scale of a model version when it
is redeployed, to match the current scale it is at. This is done using
the [Knative Initial
Scale](https://knative.dev/docs/serving/autoscaling/scale-bounds/#initial-scale)
annotation, as the Predictor / Transformer specs do not directly support
it. One caveat is that annotations applied to inference services will be
applied to all components (predictor and transformer) and thus, it is
not possible to individually control their initial scales (see related
issue: https://github.com/kserve/kserve/issues/666). Thus, the max of
the 2 values (current scale of predictor, and transformer, if enabled),
will be used as the initial scale. Autoscaling will eventually correct
the replicas as needed, and within the max threshold for the specific
component.

Note that this change is only applicable to **serverless deployments**.

Main changes:
* `api/cluster/controller.go` - Add method `GetCurrentDeploymentScale`
to the controller interface, to retrieve the current scale from the
Knative revision. If patching an existing deployment that is serverless,
get the current scale of the components and pass it down to the
templater.
* `api/cluster/resource/templater.go` - `PatchInferenceServiceSpec` now
takes in the current replicas of the existing inference service
deployment and applies the relevant annotation.

Related Turing PR: https://github.com/caraml-dev/turing/pull/350

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
When re-deploying a model version using serverless deployment, its current scale will be applied at start up.
```

**Checklist**

- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes