Enable model redeployments with SDK (#455)
**What this PR does / why we need it**:
This PR introduces a minor tweak the the SDK which allows the
redeployment of existing Merlin model versions using the `deploy`
method.

Previously, performing the following operations below will result in the
second call to `deploy` returning a `{"error":"There is 'running'
deployment for the model version"}` error:
```python
    # Create new model version
    with merlin.new_model_version() as v:
        merlin.log_model(model_dir=model_dir)

    # Deploy 
    initial_endpoint = merlin.deploy(v)

    # Deploy using raw_deployment
    new_endpoint = merlin.deploy(v, deployment_mode=DeploymentMode.RAW_DEPLOYMENT)
```

In the new implementation introduced by this PR, the second `deploy`
call will simply redeploy the router version (represented by `v`) with
the new configuration (i.e. with the raw deployment setting).

### Unrelated Modifications
This PR also introduces an additional check in the Update Endpoint
controller to ensure that an endpoint being updated is not in the
`pending` state (attempting to do so will return a 400 error). This
changes the current behaviour of this endpoint (the Update Endpoint
endpoint 🥴), as you might notice from some of the changes made to the
unit tests.

**Which issue(s) this PR fixes**:
Fixes #

**Does this PR introduce a user-facing change?**:
<!--
If no, just write "NONE" in the release-note block below.
If yes, a release note is required. Enter your extended release note in
the block below.
If the PR requires additional action from users switching to the new
release, include the string "action required".

For more information about release notes, see kubernetes' guide here:
http://git.k8s.io/community/contributors/guide/release-notes.md
-->

```release-note
1. Calling `ModelVersion.deploy` with an already deployed model version passed as a parameter will redeploy the endpoint with the new configuration of the given version
2. Updating an endpoint that is in the pending state is no longer allowed
```

**Checklist**

- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes