fix(sdk): Fix SDK model redeployment (#466)
**What this PR does / why we need it**:
This PR introduces a fix to PR #455, which itself introduced a new
feature to allow the redeployment of existing Merlin model versions
using the `deploy` method.

While this new feature assumes that a user would be able to only
**_specify fields he/she intends to update_** in the model version
configuration when calling the `deploy` method, the existing
implementation however unexpectedly overwrites some of the existing
configuration with certain default configuration values if they are not
explicitly specified. These configuration values include:
- protocol
- deployment mode
- resources
- autoscaling policy

This happens because the 'deploy' method indirectly sets default values
for the aforementioned configuration if they aren't specified (this was
primarily intended to make the **initial** deployment process simpler so
users wouldn't have to specify too many config values), and does not
distinguish between a user calling `deploy` on an existing model version
or on a new model version.

This PR thus makes changes to the SDK to set these defaults selectively.

A separate change is made to the `override` method used in the Merlin
API server update model version endpoint to override the existing model
protocol ONLY when the new model version explicitly specifies a
protocol. This change is a **_partial reversion_** of PR #300 (change 2)
which means that from now on users need to explicitly specify the HTTP
protocol (as opposed to leaving it as `null`) when updating a UPI model
to use HTTP instead. It's ultimately a minor change since the main fix
that had been introduced in that PR involves change 3 (in that PR).

**Which issue(s) this PR fixes**:
Fixes #455

**Does this PR introduce a user-facing change?**:
```release-note
Users need to explicitly specify the HTTP protocol (as opposed to leaving it as `null`) when updating a UPI model to use HTTP instead
```

**Checklist**

- [x] Added unit test, integration, and/or e2e tests
- [x] Tested locally
- [ ] Updated documentation
- [ ] Update Swagger spec if the PR introduce API changes
- [ ] Regenerated Golang and Python client if the PR introduce API
changes